fix: Fix breaking test
Tests were depending on tests from osm-p2p-geojson, and were failing because those tests use devDependencies, which are not installed for a dep.

Solution: Make a copy of test data in this repo. Depending on osm-p2p-geojson tests could easily cause breakages in the future if osm-p2p-geojson changes tests